backend/requirements.txt
Flask==2.3.2
SQLAlchemy==2.0.9
PyMySQL==1.0.2
python-dotenv==1.0.0
cryptography==41.0.3
